Generate HTML guide using tailoring file
If we use a tailoring file we can't apply XSLT directly to source
datastream file or to source XCCDF file, because we need to
resolve tailoring and add the custom profile.

The way that I propose is to load the XCCDF session first, similar
way as we do in `xccdf generate fix module`. That will resolve
resolve the tailoring automatically. Then export XCCDF benchmark
into DOM and apply XSLT on the generated tree in-memory.

The problem is that the current design doesn't add the tailoring
profiles to xccdf_benchmark profiles, but there is a separate hash
table. It would require more refactoring.

I have discovered that --hide-profile-info option doesn't do anything
because the functionality was removed 6 years ago, so I changed
documentation for it.

Resolves: RHBZ#1743835