xccdf: Introduce 'urn:xccdf:fix:script:kubernetes' fix type
This introduces the `urn:xccdf:fix:script:kubernetes` fix type which
will get added to the `system` entry in the `xccdf:fix` element. Similar
to the `urn:xccdf:fix:script:ignition`, this is initially meant for
OpenShift, however, the intent is to make these fixes more general and
applicable to other Kubernetes distributions.

This is a superset of the `urn:xccdf:fix:script:ignition` type, so it'll
eventually replace it.