utils/oscap-podman: force removal of temporary container
In some cases there are errors printed by `oscap-podman` script, it
reports about failure to remove certain paths related to a container
that is spawned by it. This seems as a race condition in `oscap-podman`
where the `podman-rm` is called right after the `podman-umount` - this
might cause that system is not fully finished with unmounting and we
request it to remove the container without waiting for the unmount of
its filesystem. This commit removes `podman-umount` command and adds
`-f` option to `podman-rm` which should take care about both, unmounting
and removing of the temporary container.