Expand loop
There is a `for` loop that runs always twice: first for
`tag = RPMTAG_BASENAMES` and second for `tag = RPMTAG_DIRNAMES`.
Instead of the `for` loop we can extract the code from inside loop and
call the function twice. We need to change the name of the function
to describe the changed operation.