fsdev: Remove undeclared variable on FreeBSD and macOS.
The FreeBSD and macOS variant of __fsdev_init had an undeclared
"fs". Fix this by removing the variable.

This issue was introduced with commits abf6adb46 and 82fee23fa.

The first removed "fs" as a function parameter, but defined it within
the body of __fsdev_init(). The second commit removed the declaration
of fs but not its later usage.